Remove roam files from agenda files
Apparently emacs autosaves frequently enough that having a crapload of open
buffers noticably impacts performance. For me, this was most apparent in
note taking in a daily roam file, where every ~5 or so seconds the notetaking buffer
would freeze. Input was still captured, but it's jarring enough that it disrupts
my note taking flow.

I'm writing a huge commit message for once because this is kind of a crappy fix - notably,
I won't be able to get my roam files in memory for use with helm-org-rifle, which will
occasionally suck. I don't use rifle frequently enough to fix this now, but when I need
it it'll be annoying. This will also have the side effect of TODOs in my roam files
not showing in my agenda.

A couple things I should probably do:
1. Fine tune auto save - https://www.gnu.org/software/emacs/manual/html_node/emacs/Auto-Save-Control.html
2. Make another keybind to load all my roam files into buffers so I can rifle them.